refactor: lecture assistant tool
