--- app/forms.py
@@ -1,8 +1,0 @@
-from flask_wtf import FlaskForm
-from wtforms.fields.simple import StringField
-from wtforms.validators import DataRequired, Length
-
-
-class LectureForm(FlaskForm):
-    question_query = StringField('Press Enter to clear.', validators=[DataRequired(), Length(max=25)],
-                                 id='search-bar')
