+++ app/templates/base.html
@@ -30,7 +30,7 @@
             <a href="{{ url_for('tools.tea_helper') }}">Tea Helper</a>
             <a href="{{ url_for('tools.item_helper') }}">Item Helper</a>
             <a href="{{ url_for('tools.seed_simulator') }}">Seed Simulator</a>
-            <a href="{{ url_for('lecture_assistant') }}">Lecture Assistant</a>
+            <a href="{{ url_for('tools.lecture_assistant') }}">Lecture Assistant</a>
         </div>
         <hr>
         <button id="theme-toggle" onclick="toggleTheme()">Toggle Dark Mode</button>
